do not make metrics its own component
it is going to be difficult to resolve source ids under one endpoint

e.g. /metrics/:guid/cpu-aggregated-series

so instead we will keep everything under the resources individually

e.g. /applications/:guid/metrics -> return all metrics for an app
e.g. /applications/:guid/metrics/cpu-aggregated-series -> returns series
