+++ src/components/service-metrics/service-metrics.ts
@@ -1,19 +1,19 @@
 import moment from 'moment-timezone';
 
 import CloudFoundryClient from '../../lib/cf';
-import { timeOffsets } from '../../lib/metrics';
-import PromClient from '../../lib/prom';
-import { IParameters, IResponse } from '../../lib/router';
-
-import { IContext } from '../app/context';
-import { IBreadcrumb } from '../breadcrumbs';
-
 import {
   rdsCPUUsageAggregatedSeries,
 
   rdsFreeStorageSpaceAggregatedSeries,
   rdsFreeStorageSpaceSingleStat,
-} from '../metrics';
+
+  timeOffsets,
+} from '../../lib/metrics';
+import PromClient from '../../lib/prom';
+import { IParameters, IResponse } from '../../lib/router';
+
+import { IContext } from '../app/context';
+import { IBreadcrumb } from '../breadcrumbs';
 
 import serviceMetricsTemplate from './service-metrics.njk';
